feat: add ability to render react node in AgentMessage
