+++ packages/chat/src/components/LoadingDots/LoadingDots.story.tsx
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
+import AgentMessage from '../AgentMessage';
 import { LoadingDots } from '.';
 
 type Story = StoryObj<typeof LoadingDots>;
@@ -12,3 +13,15 @@ const meta: Meta<typeof LoadingDots> = {
 export default meta;
 
 export const Base: Story = {};
+
+export const InMessage: Story = {
+  render: () => {
+    return (
+      <div>
+        <AgentMessage from="system">
+          <LoadingDots />
+        </AgentMessage>
+      </div>
+    );
+  },
+};
